backend/requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
pyqt5
torch
torchvision
pydantic
requests